moved ApplicationParams class under android utils package
