+++ jadx-gui/src/main/java/jadx/gui/ui/MainWindow.java
@@ -86,13 +86,13 @@ import jadx.api.ResourceFile;
 import jadx.api.plugins.events.IJadxEvents;
 import jadx.api.plugins.utils.CommonFileUtils;
 import jadx.core.Jadx;
-import jadx.core.export.ApplicationParams;
 import jadx.core.export.TemplateFile;
 import jadx.core.plugins.events.JadxEventsImpl;
 import jadx.core.utils.ListUtils;
 import jadx.core.utils.StringUtils;
 import jadx.core.utils.android.AndroidManifestParser;
 import jadx.core.utils.android.AppAttribute;
+import jadx.core.utils.android.ApplicationParams;
 import jadx.core.utils.exceptions.JadxRuntimeException;
 import jadx.core.utils.files.FileUtils;
 import jadx.gui.JadxWrapper;
